// src/lib/model/weights.ts
export const MODEL_DESCRIPTOR: ModelDescriptor = {
  "featureCount": 558,
  "defaultThreshold": 0.45,
  "modelType": "Autoencoder",
  "projectName": "SECOM Anomaly Detection",
  "kerasVersion": "3.10.0",
  "metrics": {
    "precisionAnomaly": 0.4457831325301205,
    "recallAnomaly": 0.3557692307692308,
    "f1Anomaly": 0.39572192513368987,
    "accuracy": 0.7153652392947103
  },
  "layers": [
    {
      "name": "dense",
      "inputSize": 558,
      "outputSize": 128,
      "activation": "relu"
    },
    {
      "name": "dense_1",
      "inputSize": 128,
      "outputSize": 64,
      "activation": "relu"
    },
    {
      "name": "dense_2",
      "inputSize": 64,
      "outputSize": 32,
      "activation": "relu"
    },
    {
      "name": "dense_3",
      "inputSize": 32,
      "outputSize": 64,
      "activation": "relu"
    },
    {
      "name": "dense_4",
      "inputSize": 64,
      "outputSize": 128,
      "activation": "relu"
    },
    {
      "name": "dense_5",
      "inputSize": 128,
      "outputSize": 558,
      "activation": "linear"
    }
  ]
} as const;